pyproject.toml
[project]
name = "coffee-shop-api"
version = "0.1.0"
description = ""
authors = [
    { name = "Aleksandr", email = "alex.tawila@gmail.com" }
]
readme = "README.md"
requires-python = "^3.12"

dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "celery[redis] (>=5.5.3,<6.0.0)",
    "aiosmtplib (>=4.0.1,<5.0.0)",
    "pydantic-settings (>=2.10.0,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)"
]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.16.1"
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"

# здесь — массив таблиц вместо одиночного блока
[[tool.poetry.packages]]
include = "app"
from = "."

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"